+++ package.json
@@ -12,6 +12,7 @@
     "@types/react": "^18.0.21",
     "@types/react-dom": "^18.0.9",
     "@ya.praktikum/react-developer-burger-ui-components": "^1.13.4",
+    "history": "^5.3.0",
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-dnd": "^16.0.1",
@@ -20,6 +21,7 @@
     "react-redux": "^8.0.4",
     "react-router": "^6.4.1",
     "react-router-dom": "^5.2.1",
+    "react-router-redux": "^4.0.8",
     "react-scripts": "5.0.1",
     "react-transition-group": "^4.4.5",
     "redux": "^4.2.0",
