+++ src/components/ForgotPassForm/ForgotPassForm.js
@@ -9,38 +9,18 @@ import FormNavigationElement from "../FormNavigationElement/FormNavigationElemen
 import { useForm } from "../../hooks/useForm";
 import { useDispatch, useSelector } from "react-redux";
 import { useHistory } from "react-router-dom";
-
-import { FORGOT_PASS_URL } from "../../utils/constants";
-import { api } from "../../utils/api";
-import {
-  ERROR_FORGOT_PASS,
-  START_FORGOT_PASS,
-  SUCCESS_FORGOT_PASS,
-} from "../../services/actions/forgot-pass";
+import { restorePassword } from "../../services/asyncActions/restorePassword";
 
 const ForgotPassForm = () => {
   const dispatch = useDispatch();
   const { values, handleChange, isValid } = useForm();
-  const { fetchPost } = api(FORGOT_PASS_URL);
+
   const history = useHistory();
   const { isLogged } = useSelector((store) => store.login);
 
-  const postRegistrationDetails = (data) => {
-    dispatch({ type: START_FORGOT_PASS });
-    fetchPost(data)
-      .then((response) => {
-        if (response && response.success) {
-          dispatch({ type: SUCCESS_FORGOT_PASS, payload: response });
-          dispatch({ type: ERROR_FORGOT_PASS });
-          history.push("/reset-password");
-        }
-      })
-      .catch((err) => dispatch({ type: ERROR_FORGOT_PASS }));
-  };
-
   const handleSubmit = (e) => {
     e.preventDefault();
-    postRegistrationDetails(values);
+    dispatch(restorePassword(values));
   };
 
   useEffect(() => {
