+++ src/utils/constants.js
@@ -1,3 +1,4 @@
+import { createBrowserHistory } from "history";
 import PropTypes from "prop-types";
 
 const BASE_URL = "https://norma.nomoreparties.space/";
@@ -17,7 +18,7 @@ export const TYPE_MAIN = "main";
 
 export const reactModalRootElement = document.querySelector(".react-modals");
 
-export const modalRoot = document.getElementById("react-modals");
+export const browserHistory = createBrowserHistory();
 
 export const INGREDIENT_TYPES = PropTypes.shape({
   _id: PropTypes.string.isRequired,
